+++ tests/testthat/test_check_df_visit.R
@@ -1,5 +1,5 @@
 
-df_visit <- get_df_visit_test()
+df_visit <- get_df_visit_test_mapped()
 
 test_that("check_df_visit() must fix missing visits and throw a warning", {
   df_visit_filt <- df_visit %>%
@@ -25,7 +25,7 @@ test_that("check_df_visit() must aggregate duplicated visits and throw a warning
 test_that("check_df_visit() must thrown an error when n_ae and visit columns are not numeric", {
   expect_error({
     df_visit %>%
-      mutate_at(vars(n_ae, visit), as.character) %>%
+      mutate_at(vars(n_event, visit), as.character) %>%
       check_df_visit()
   },
   regexp = "event number and visit columns must be numeric"
@@ -39,7 +39,7 @@ test_that("check_df_visit() must thrown an error when NA's are detected", {
       bind_rows(df_visit) %>%
       check_df_visit()
   },
-  regexp = "NA detected in columns: study_id,site_number,patnum,n_ae,visit"
+  regexp = "NA detected in columns: study_id, site_number, patnum, n_event, visit"
   )
 })