+++ tests/testthat/test_data.R
@@ -1,4 +1,5 @@
-df_visit <- get_df_visit_test()
+df_visit <- get_df_visit_test_mapped()
+
 df_portf_test <- get_df_portf_test()
 
 test_that("sim_ur() ae_count as expected for max patient visits", {
@@ -6,14 +7,14 @@ test_that("sim_ur() ae_count as expected for max patient visits", {
   n_ae_test <- df_visit %>%
     filter(visit == max(visit), .by = c("study_id", "site_number", "patnum")) %>%
     filter(site_number == "S0001", study_id == "A") %>%
-    pull(n_ae) %>%
+    pull(n_event) %>%
     sum()
 
   n_ae_0p5 <- df_visit %>%
     sim_ur(study_id = "A", site_number = "S0001", ur_rate = 0.5) %>%
     filter(visit == max(visit), .by = c("study_id", "site_number", "patnum")) %>%
     filter(site_number == "S0001", study_id == "A") %>%
-    pull(n_ae) %>%
+    pull(n_event) %>%
     sum()
 
   expect_true(n_ae_test == n_ae_0p5 * 2)
@@ -143,36 +144,73 @@ test_that("sim_test_data_portfolio() produces the expected output when ae_rates 
   expect_true(any(! unique(ae_rates_test_sd[["stdev"]]) == 0))
 })
 
-test_that("sim_test_data_study() alters the ae_rates value if is_ur is TRUE", {
-  expect_true(unique(sim_test_data_study(frac_site_with_ur = 2,
-                                         ae_rates = 2, ur_rate = 1)[["ae_per_visit_mean"]]) == 0)
+test_that("sim_test_data_study() alters the ae_rates value if ratio_out > 0", {
+
+  n_unique_ae_rates <- sim_test_data_study(
+      ratio_out = 0,
+      event_names = "ae"
+    ) %>%
+    pull(.data$ae_per_visit_mean) %>%
+    unique() %>%
+    length()
+
+  expect_true(n_unique_ae_rates == 1)
+
+  n_unique_ae_rates <- sim_test_data_study(
+      ratio_out = 0.3,
+      factor_event_rate = -0.5,
+      event_names = "ae"
+    ) %>%
+    pull(.data$ae_per_visit_mean) %>%
+    unique() %>%
+    length()
+
+  expect_true(n_unique_ae_rates == 2)
+
 })
 
 test_that("purrr_bar() - .slow is TRUE", {
+
   param <- (rep(0.25, 5))
   purr_test <- purrr_bar(param, .purrr = purrr::walk, .f = Sys.sleep, .steps = 5, .slow = TRUE)
   purr_test2 <- purrr_bar(param, .purrr = purrr::walk, .f = Sys.sleep, .steps = 5, .slow = FALSE)
   expect_equal(purr_test, purr_test2)
-  })
 
-test_that("sim_test_data_study() produces an error
-          when the number of event names != the number of events per visit means", {
-          expect_error(sim_test_data_study(event_names = c("ae", "pd"), ae_per_visit_mean = 0.5),
-regexp = "Number of events named (2) doesn't equal the number of events per visit means submitted (1)",
-                       fixed = TRUE)
-          })
-
-test_that("sim_test_data_study() produces an error
-          when the number of event names != the number of event rates submitted as a list", {
-          expect_error(sim_test_data_study(event_names = "ae", ae_rates = list(0.3, 0.4)),
-                       regexp = "Number of events named (1) doesn't equal the number of events rates submitted (2)",
-                       fixed = TRUE)
 })
 
-test_that("sim_test_data_study() produces an error
-          when the number of event names > 1 and event rates are submitted as a vector", {
-  expect_error(sim_test_data_study(event_names = c("ae", "pd"),
-                                   ae_per_visit_mean = c(0.4, 0.4), ae_rates = c(0.3, 0.4)),
-               regexp = "ae_rates should be entered as a list (containing arrays) when the number of events is > 1",
-               fixed = TRUE)
+test_that(paste("sim_test_data_study() produces an error",
+          "when the number of event names != the number of events per visit means"), {
+
+  expect_error(
+    sim_test_data_study(event_names = c("ae", "pd"), event_per_visit_mean = 0.5),
+    regexp = "Number of events named (2) doesn't equal the number of events per visit means submitted (1)",
+    fixed = TRUE
+  )
+
+})
+
+test_that(paste("sim_test_data_study() produces an error",
+          "when the number of event names != the number of event rates submitted as a list"), {
+
+  expect_error(
+    sim_test_data_study(event_names = "ae", event_rates = list(0.3, 0.4)),
+    regexp = "Number of events named (1) doesn't equal the number of events rates submitted (2)",
+    fixed = TRUE
+  )
+
+})
+
+test_that(paste("sim_test_data_study() produces an error",
+          "when the number of event names > 1 and event rates are submitted as a vector"), {
+
+  expect_error(
+    sim_test_data_study(
+      event_names = c("ae", "pd"),
+      event_per_visit_mean = c(0.4, 0.4),
+      event_rates = c(0.3, 0.4)
+    ),
+    regexp = "event_rates should be entered as a list (containing arrays) when the number of events is > 1",
+    fixed = TRUE
+  )
+
 })